refactor: renomme HelpButton en InfoBulle et supprime ancien composant du même nom
