+++ site/source/design-system/index.ts
@@ -20,7 +20,7 @@ export { TitreObjectifSaisissable } from './typography/TitreObjectifSaisissable'
 export * from './layout'
 
 export * from './buttons'
-export { HelpButton } from './molecules/HelpButton'
+export { InfoBulle } from './molecules/InfoBulle'
 export { InfoButton } from './molecules/InfoButton'
 export * from './card'
 export * from './accordion'
@@ -31,7 +31,6 @@ export * from './conversation'
 export * from './molecules/field'
 export * from './drawer'
 export * from './emoji'
-export { InfoBulle } from './InfoBulle'
 export * from './message'
 export * from './footer'
 export { Popover } from './popover/Popover'
